# server/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "OpinionLK API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3002",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/signup": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/auth/getmembers": {
      "get": {
        "tags": [
          "Community Manager"
        ],
        "description": "Endpoint to get all community managers",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/savemember": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ManagerFirstName": {
                  "example": "any"
                },
                "ManagerLastName": {
                  "example": "any"
                },
                "ManagerAddLine1": {
                  "example": "any"
                },
                "ManagerAddLine2": {
                  "example": "any"
                },
                "ManagerDistrict": {
                  "example": "any"
                },
                "ManagerEmail": {
                  "example": "any"
                },
                "ManagerPhone": {
                  "example": "any"
                },
                "ManagerNic": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/auth/updatemember/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ManagerFirstName": {
                  "example": "any"
                },
                "ManagerLastName": {
                  "example": "any"
                },
                "ManagerAddLine1": {
                  "example": "any"
                },
                "ManagerAddLine2": {
                  "example": "any"
                },
                "ManagerDistrict": {
                  "example": "any"
                },
                "ManagerEmail": {
                  "example": "any"
                },
                "ManagerPhone": {
                  "example": "any"
                },
                "ManagerNic": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/auth/deletemember/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/client/signup": {
      "post": {
        "tags": [
          "Organisation"
        ],
        "description": "Endpoint to sign up organisation",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "orgName": {
                  "example": "any"
                },
                "orgAddressLine1": {
                  "example": "any"
                },
                "orgAddressLine2": {
                  "example": "any"
                },
                "orgCity": {
                  "example": "any"
                },
                "orgState": {
                  "example": "any"
                },
                "orgZip": {
                  "example": "any"
                },
                "orgPhone": {
                  "example": "any"
                },
                "orgEmail": {
                  "example": "any"
                },
                "orgWebsite": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "position": {
                  "example": "any"
                },
                "department": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "nic": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "confirmPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/client/clientdata": {
      "get": {
        "tags": [
          "Organisation"
        ],
        "description": "Endpoint to get organisation data",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/allsurveys": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/user/userdata": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get user data",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/survey/createResponse": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "surveyid": {
                  "example": "any"
                },
                "response": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/survey/getbyUserId": {
      "get": {
        "tags": [
          "Organisation",
          "Community Manager"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/survey/getbySurveyId/{surveyid}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "surveyid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/survey/all": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/survey/getsurvey/{surveyid}": {
      "get": {
        "tags": [
          "Organisation",
          "Community Manager"
        ],
        "description": "Gets survey to edit, checks edit privileges",
        "parameters": [
          {
            "name": "surveyid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/survey/editquestion/{surveyid}/{questionid}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "surveyid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/survey/getQuestion/{surveyid}/{questionid}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "surveyid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/survey/addQuestion/{surveyid}": {
      "post": {
        "tags": [
          "Organisation",
          "Community Manager"
        ],
        "description": "",
        "parameters": [
          {
            "name": "surveyid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/survey/deleteQuestion/{surveyid}": {
      "put": {
        "tags": [
          "Organisation",
          "Community Manager"
        ],
        "description": "",
        "parameters": [
          {
            "name": "surveyid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "questionid": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/survey/create": {
      "post": {
        "tags": [
          "Organisation",
          "Community Manager"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/survey/changestatus/{surveyid}": {
      "put": {
        "tags": [
          "Organisation",
          "Community Manager"
        ],
        "description": "",
        "parameters": [
          {
            "name": "surveyid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "state": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/survey/imageUpload": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}